requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.5.0
numpy>=1.24.0
runpod>=0.9.0
matplotlib>=3.5.0
tiktoken>=0.5.0
einops>=0.6.0
transformers_stream_generator>=0.0.5
accelerate>=1.0.0
